organizing and refactoring
