+++ agent/system/structs.go
@@ -1,0 +1,13 @@
+package system
+
+import "time"
+
+type CmdOptions struct {
+	Shell        string
+	Command      string
+	Args         []string
+	Timeout      time.Duration
+	IsScript     bool
+	IsExecutable bool
+	Detached     bool
+}
